Switch back to old CpuId implementation
Like several KVM-based structures, the CpuId structure contains an
underlying zero-sized array. On the rust-vmm side, someone was working
on a Vector implementation to abstract this more safely, and this safer
abstraction was temporarily inserted into the vmm-vcpu branch used
here. Switch back to the old/original CpuId implementation to be
compatible with crosvm code.