Add InterruptEvent struct
This is the first change out of the few will have to make in order
to be able to emulate KVM's API.

KVM allows mapping eventfds to IRQs, which will be triggered when
the events get signaled.

Since WHP doesn't support this, we'll need to do it ourselves. This
requires an additional structure that mimics the EventFd interface,
triggering the requested interrupt when signaled. For Linux, we'll
just typedef it to EventFd.

For convenience, this will also act as an EventFd wrapper so that
it can easily be used for resample events as well (which will also
have to be emulated in a subsequent patch).