Pull vmm-vcpu whp implementation into the "whp" crate
We're moving this abstraction layer into the "whp" crate, which
may become a rust-vmm crate. This will allow us to more easily
extend this abstraction, emulating some of the KVM api
(e.g. ioeventfd, resample events, etc).

At the same time, this abstraction layer was a bit out of the scope
of libwhp.