Use LapicStateRaw instead of vmm-vcu LapicState
This will allow us to drop the libwhp vmm-vcpu dependency and move
forward with upstreaming the libwhp changes.

We assume that the two structures are identical, but this cast is
ugly/unsafe. Since vmm-vcpu is a compatibility layer, it's
probably better if we re-export the libwhp structure there and have
vmm-vcpu depend on libwhp instead of the other way arond.